scripts/requirements.txt
discord.py[voice]>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
yt-dlp>=2023.12.0
PyNaCl>=1.5.0
openai>=1.0.0
